Moved License model to separate folder
