+++ FOSSDB_web/apps/fossdb/models.py
@@ -3,17 +3,9 @@ import uuid
 from django.conf import settings
 from django.db import models
 
+from .license.models import License
+
 User = settings.AUTH_USER_MODEL
-
-
-class License(models.Model):
-    short_name = models.CharField(max_length=50)
-    full_name = models.CharField(max_length=100, blank=True, default="")
-    url = models.URLField(blank=True, default="")
-    description = models.TextField(blank=True, default="")
-
-    def __str__(self):
-        return self.short_name
 
 
 class HostingPlatform(models.Model):
